Add bfrange support
The ToUnicode CMap can express the correspondence
between CID/GID and Unicode in two types: bfchar and bfrange.
In most cases, bfchar is used when the correspondence is discrete,
while bfrange is used when the correspondence is continuous.
For CJK characters, bfchar is used in most cases
because their correspondence is discrete.
Threfore, pdf-fix-tuc did not support bfrange,
but some PDFs use bfrange for CJK characters
even though the correspondence is discrete.

This commit adds bfrange support.
It makes pdf-fix-tuc can handle PDFs
which represent ToUnicode CMaps in bfrange.